plugins/cni: Remove EP configuration hook
As of cilium/cilium#42999, we now have an alternative approach to
customizing the CNI ADD behavior. This commit removes the old EP
configuration version.
